move some logic into CustomizePanel
